// redux/types.ts
export enum types {
    setStickers = 'SET_STICKERS',
    removeSticker = 'REMOVE_STICKER',


    addStickerToSet = 'ADD_STICKER_TO_SET',
    removeStickerFromSet = 'REMOVE_STICKER_FROM_SET',
    resetStickers = 'RESET_STICKERS',
    setQuickEditSticker = 'SET_QUICK_EDIT_STICKER',
    setEditSticker = 'SET_EDIT_STICKER',
    setSubjects = 'SET_SUBJECTS',




    setUser = 'SET_USER'
}